Implement migration support
Migrations are stored under db/migrations with the format
N_descripive_name_for_migration.sql. N will be passed as an integer and
used for ascending ordering, so migrations with lower numbers are
applied first.

The migration system keeps track of which migrations have been
successfully applied. A backup of the database is made prior to applying
any migrations, the backup will be alongside the existing database but
with a timestamp in the name. If the backup cannot be generated then no
migrations are applied. Migration files should consist of zero or more
valid SQL statements. Each migration is applied in a separate
transaction. Within each transaction an additional row will be inserted
into the migrations table to indicate successful application of the
migration. If any migration generates an error, no more migrations are
applied.

If the database is empty then a fast forward path is followed and
db/latest_schema.sql is applied.