Use the parent context on operations in DB txns
At first I didn't use the sql.Tx.*Context operations because I was
confused why a transaction opened with a context would allow a different
context on the inner operations. After reading the implementation it
became clear that the *Context methods only check in the provided
context is Done. And the non-context versions would internally call the
corresponding Context method with context.Background, which cannot be
cancelled.

So it makes sense to use the *Context methods and provide the same
context as used to open the transaction.