Fix a bug in the migration fast forward logic
Previously if the DB was determined to be empty then a fast forward was
applied which directly applies the latest schema to the DB. This is fine
but it did not update the migrations table, meaning that on next
application start the migration logic would attempt to apply the entire
set of migrations which will fail because the DB is already "at latest".

The fix is to populate the migrations table on a fast forward. This does
make the reasonable assumption that the latest schema is equal to
applying all the migrations. This is on the developer to ensure.