Start to implement search queries
This works by computing an embedding vector for the query and then
iterating over all description embedding vectors and finding those with
the best cosine similarity. This is O(N) for all embeddings because
there is no easy way to accelerate the search on embedding vectors.

Make .gitignore rule more specific to a single file, so it doesn't
cover paths that contain henri.