Complete EmbeddingsForModel
Question about whether we should be exposing the channels to the caller.
Would it be better to have EmbeddingsForModel return all the data but
internally use go routines to query batches of data from the DB
concurrently? Feels like we are making things unnecessarily complicated
for the caller, this pattern allows the caller to terminate the query
midway through, but in this case we want all the results anyway.